Updated openstack/openstack Project: openstack/oslo.db  bdcfb4a4df4346e7e05473a7579b1da40212bf9b
Let oslotest manage the six.move setting for mox

This instruction to set up a move in six for the mox module conflicts
with the one in oslotest. We are trying to use mox3 everywhere, so
remove this instruction and let oslotest handle it.

Imported from the incubator change with the same ID.

Change-Id: I59d5799283233f8411044ddb15c8abfc8850014c