Updated openstack/openstack Project: openstack/nova  4b8ad7ae80c37bf58785973278c360e642005b8b
Ironic driver backports: patch 7

The patch to remove the dependency on python-ironicclient was
misplaced in the split of 103167 into the series of smaller
patches that we're now reviewing.  This patch reinstates this
change so ironicclient is loaded dynamically, and does not
become a requirement for Nova when not using Ironic.

Change-Id: I70a2fa8fc45e18524ea4d777d72da24d5d329c25
Co-authored-by: Devananda van der Veen <devananda.vdv@gmail.com>