Updated openstack/openstack Project: openstack/oslo.db  8f7681946a62b5f835f267b68d02bf4c8aa7a0e2
Renaming in WalkVersionsMixin

Rename _walk_versions(), _migrate_up(), _migrate_down()
and mark them deprecated as they will be removed in few
releases. These functions will be the part of public API
as this class will be used for migrations in other
OpenStack projects, so there is no need to make them
'private'.

Change-Id: Ic9358445e60a0dd43a5900e8bda7b12f2bebf679