Updated openstack/openstack Project: openstack/python-keystoneclient  0823039e30dfb46e63f4d1533982b47fa333c940
Allow providing an endpoint_override to requests

As much as I'd prefer not to need this functionality there are plenty of
existing clients that we want to have use the adapter that can accept a
bypass argument such that it ignores the service catalog and uses that
URL for all requests.

We therefore need to be able to support similar functionality in our
adapter.

Change-Id: I206705241ff9b84967d0d9c089b4795bcc26b65e