Updated openstack/openstack Project: openstack/python-keystoneclient  a9367009ad3ab1da003bb5e43d1ae1e3ecfee2a3
Add version parameter to adapter.

The version parameter was left out of adapter previously because setting
a version number triggers discovery and I felt that it was not ready for
the other services.

However if it isn't set then it won't be used and we may as well
implement it here once rather than have the individual services that do
support it override it themselves.

Change-Id: I707380a01175dc19b59de32cbb8fd2bb123d7335