Updated openstack/openstack Project: openstack/ceilometer  261c61b63ed6e1cecc8353c88a392e7873454ed5
Migrate the rest of the central agent pollsters to use discoveries

Currently, not all the central agent pollsters use discoveries.
In order to enable horizontal scaling, all pollsters must use
discoveries so the resources can be partitioned across the pool
of running central agents.

For pollsters that don't poll specific resources, but only poll
general info from services, we treat the service endpoints as
resources. So in the case of Glance, there will be only one
resource available for all the pollsters (a Glance endpoint).
If there are multiple agents running, only one will be assigned
that endpoint and only one will poll Glance API.

DocImpact
Co-Authored-By: Dina Belova <dbelova@mirantis.com>
Closes-Bug: #1364352
Change-Id: I8f3b228db9aacf3a7cc4b719c50013cc30d5aa79