Updated openstack/openstack Project: openstack/neutron  5e841857b685daa3b564b0ba8f519fe725211bb6
ofagent: Remove @author tags and update copyright notices

When writing these modules, ofagent team (kakuma, yamamoto) thought
that @author tag was OpenStack's style to represent authorships.
But it turned out to be wrong.  Acutally there seems to be a consensus
not to use them.  So, replace them back to the original Ryu-style
copyright notices.

Change-Id: Icc9fdd1850c0c1dbd354bb3312199a7f6849925a