Updated openstack/openstack Project: openstack/cinder  dbe937f8127b2268bc8fb267eb152e414eff6e0a
Set default pool value to system in gpfs driver

The configuration option gpfs_storage_pool is assigned with
default value of None. And it gets initialized with None
instead of system, where as the intention is to set the default
value as system.

This change sets default value of gpfs_storage_pool to system
instead of None.

DocImpact

Needs an update in gpfs driver documentation, as this patch
proposes a change in default value of configuration option
(gpfs_storage_pool)

Change-Id: Ib4c2a2e4593a940ae563d329fd59d2d5fe3d4520
Closes-Bug: #1365474