Updated openstack/openstack Project: openstack/cinder  69b4cfd09c61da1ccd9c9587dcd043240bc3e9d0
Fixes Cinder fails to upload volume to vhd image

The method copy_volume_to_image from windows.py specifies the
wrong volume format, namely 'vpc'. In this case, the upload_volume
method from image_utils will attempt to convert the volume to vhd
which will result in an error from qemu as it does not recognize
the format.

This patch fixes this issue by using 'vhd' as a format when
uploading volumes

Change-Id: Ifa675316e19f9a795e103fd56fcdf4e3ed00aab7
Closes-Bug: #1365487