Updated openstack/openstack Project: openstack/nova  72b2555339e5aa24551e3171134412d8d23a8bff
Add pci_requests to instance_extra table

Add a new text column, pci_requests, to the instance_extra table.
This column will be used to hold the data that is currently stored in
instance system_metadata.  The problem with its current location is
the 255 character limit, which becomes a more serious limitation with
the coming SR-IOV support.

Related to blueprint pci-passthrough-sriov

Change-Id: I84d02242c56ad56c597fe588407aaceb24910d0a