Updated openstack/openstack Project: openstack/nova  f6340ea106d29b37d593bf0fc4819893dec78b21
Expose numa_topology to the resource tracker

This patch makes all the current drivers that implement
get_available_resource() return an additional 'numa_topology' field in
the resources dict. The libvirt driver will actually return useful data,
while other drivers only return 'None' at this time.

We also make the resource tracker check for the presence of this field.
The intent is that it be mandatory for all drivers to expose, but to
avoid breaking out of tree virt drivers, if the field is missing it
will be set to None.

Blueprint: virt-driver-numa-placement
Change-Id: I75e4aeea4636532cfd724f718e136fac5bb2d15e