Updated openstack/openstack Project: openstack/nova  c886168aae9b1bfaf253d71ddda0a31805a0705c
Add claims testing to VirtNUMAHostTopology class

This will allow us to use NUMA topology information from hosts and
instances when making resource claims for scheduling and migrating
instances.

This patch adds the VirtNUMALimitTopology class and corresponding
cells class that are used to represent the limits, and a method
claim_test() on the VirtNUMAHostTopology class that does the checking.

Blueprint: virt-driver-numa-placement
Change-Id: Ic21d559edf292e5b5dc1bba966e10ec4af0f706e