Updated openstack/openstack Project: openstack/ceilometer  c5ce0a703f3caed802a1404906d5506a01546e16
fix network discovery meters

currently we let neutronclient authenticate against keystone. this
makes it impossible for us to figure out whether service is valid
until too late (when we call poll neutronclient).

this patch adds keystoneclient to check_keystone decorator so it
can verify endpoints when keystoneclient is not available.

Closes-Bug: #1361818
Change-Id: Ia374e9b422ed8fa088b28ecf200161248136b09e