Updated openstack/openstack Project: openstack/horizon  9c0d905caa72b0f1563f50f71bf751ab947a8988
Allow adding or changing the CSS classes of a data table

Sometimes a datatable needs to be displayed slightly differently, for
instance with or without the borders, condensed, zebra, etc. This patch
lets us do that in two ways.

It adds a css_class option to table's Meta, so that we can easily add
more classes to a particular table, and modify its looks with CSS.

It also puts the whole table's CSS class list in a block, so that we
can inherit from the table's template to completely replace its classes,
if we need to.

Change-Id: I64a1f3ba8d80643c545de024f52ec3313bbff8e5