Updated openstack/openstack Project: openstack/keystone  6fbb5b82c8f5a566f61f940b2bedb0209343af5d
Add V3 JSON Home support to GET /

The server wasn't returning a JSON Home response for GET / when the
Accept header is `application/json-home`. By returning the V3 JSON
Home response for GET / a V3 client that supports JSON Home can
GET either /v3 or / and use the response. The identity API should
be able to be set to /.

Closes-Bug: #1366589
Change-Id: I3191a85acf9d2f582f6b48a164cf5ac2bf84a8cf