Updated openstack/openstack Project: openstack/keystone  797adeada505b277cdb5404ed021138297d07c6c
Make the extension docs a top level entry in the landing page

Currently any documentation about extensions that Keystone offers
is lacking at best. This patch moves the documentation out of the
busy configuration section, and to index/landing page.

It also provides a brief summary for each extension, and a logical
point of entry for additional information.

Change-Id: Ide8ac9b277e0fc9cab9c01073bb72fe3bc369aed