Updated openstack/openstack Project: openstack/oslo.config  059579ac2189b94bc9e9555b2e9acfb31a83ef53
Looks for variable subtitution in the same group

When we do a variable subtitution we must try to look if the option
is avialable into the same group and then into the default one for backward
compatibilty.

Otherwise variable subtitution doesn't works when we change the group of
an option and deprecated the DEFAULT group.

Change-Id: Ie5750ee028a58fba6da225a8c4b4221e23db829f
Closes-bug: #1367790