Updated openstack/openstack Project: openstack/neutron  7c86ec9b738c0f361fa6e3c6dd2321db2a5117e5
UTs: Disable auto deletion of ports/subnets/nets

Change the behavior of the unit test context managers
for ports, networks, and subnets to skip the deletion
of the objects they create.

Adds explicit delete calls to the tests that require
deletion of the objects.

This is the first patch in a series to eliminate the
automatic deletion of objects created during tests to
improve test performance and make testing more explicit.

Implements: blueprint remove-unit-test-autodeletion
Change-Id: Ifc4490b121310ad64e18936e6fd94a2c9a1e36f2