Updated openstack/openstack Project: openstack/ceilometer  f47161a2bb8b08016892a69c0e4b0b3d155c7ca8
Allow to pass dict from resource discovery

Allowing to pass dict from resource discovery and storing it
as metadata. Before it was only possible to pass IP address,
which is not enough.
For TripleO case, important metadata like image name and flavor
from Nova are passed and stored in resource, allowing to make
aggregate queries according to them.

Change-Id: I4dabf79f248842398e1126d310d110b1fe95fad5
Closes-Bug: #1366871