Updated openstack/openstack Project: openstack/tempest  c511d4cb91e0670ac985fbe8fcb075c883cc1ec5
Add a NotLockingAccounts credentials provider

This credentials provider does not do any locking, so it
does not suffer from the lack of credentials cleanup which
affects the regular locking Accounts class.

Using thie credentials provider allows to keep it possible
to run tempest with no tenant isolation, and in serial mode,
as it is done today by one of periodic jobs.

Partially-implements bp:test-accounts
Change-Id: I160d183a299d2e46d269428409af680c0fdb03b5