Updated openstack/openstack Project: openstack/tempest  cd2cad362bdbb72a8b929c62a2336cb8e3bd857e
Have the EC2 instance test to wait for termination

Several EC2 test did not waited for instance termination.
This change adds an explicit wait for termination,
for the test cases which wants to terminate
the instance inside the test case.

The #1365831 happened because the 'test_run_terminate_instance',
did not waited for the final termination.

Change-Id: Icd3b86ebc8464bd20319e8582541d78fded10ee2
Closes-Bug: #1365831