Updated openstack/openstack Project: openstack/tempest  45ffc546b33e75ea68028d3ac63e783fd9c83b64
Allow FWaaS API test to accept more than one end state

To support distributed routers, it is useful to have this API tempest test
to accept a firewall resource to be either in ACTIVE or CREATED state. After
all, this is not a scenario test, so whether the resoure is actually wired
should be irrelevant.

Closes-bug: #1360351

Change-Id: I2367c2877ca2ebabc08095b33a85659de2b882bd