Updated openstack/openstack Project: openstack/ceilometer  b8366ed9d4bfb14f65d7a1044558d9508fdecbe9
Sync jsonutils for namedtuple_as_object fix

This is to pull in the fix for bug 1361230 which was introduced
with a prior sync of jsonutils.

strutils is pulled in as a dependency but also fixes an
unrelated bug 1345233.

Changes:

jsonutils
---------
d9060f6 Check for namedtuple_as_object support before using it

strutils
--------
66142c3 Make strutils.mask_password more secure

Change-Id: I9bd2e5f93c2684203487627f64f70d5da7b13586
Closes-Bug: #1361230
Partial-Bug: #1345233