Updated openstack/openstack Project: openstack/nova  63a5bbaf9827d0de6c6cec0805b83c8001cba1f3
document why we disable usb_tablet in code

This provides a code level explanation for why we disregard the user
option for use_usb_tablet when the spice agent is enabled. It came up
in a bug report, and having the comment close to the code helps carry
the information forward.

Change-Id: I7945e921b6283519d91e234584d99d9e82218e6d
Closes-Bug: #1356633