Updated openstack/openstack Project: openstack/keystone  6fad5cf3b967679d75594e06f90ab0493b73a809
Use id attribute map for read-only LDAP

For read-only LDAP, we have no control over the DIT. We can only use the
attribute maps. Therefore, we need to honor the attribute map for all
attributes.

This patch also added the missing ID attribute when creating an object in
LDAP. For lookup, we also need the ID attribute to be returned.

Change-Id: I57c099eefdca7e756d294c35cf5b70d119c5c955
closes-bug: #1361306
closes-bug: #1340041