Updated openstack/openstack Project: openstack/nova  3002d4333b93c1f5923dbc1925f6f2a24b314472
Improve logging of external events on the compute node

We don't log every event we receive on the compute node, which could
allow some to slip through and look like they were never received. This
patch makes sure to debug log each one so we at least know that it made
it from API to compute.

Change-Id: I817248ea8167be21c55fecad3a39cb45b027596c
Related-bug: #1357476